requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
aiosqlite>=0.20.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
itsdangerous>=2.2.0
supabase>=2.4.0